MEDIA-57 #comment Deprecated old SDP classes #Continue wiring new SDP library

+++ io/ice/src/main/java/org/mobicents/media/io/ice/IceCandidate.java
@@ -131,7 +131,7 @@ public class IceCandidate extends TransportAddress implements Comparable<IceCand
 		return (IceCandidate) super.clone();
 	}
 
-	public int getComponentId() {
+	public short getComponentId() {
 		return this.component.getComponentId();
 	}
 }
